==> scripts/3_buffer_encounter_plots.R <==
rm(list=ls())
gc()
library(cmocean)
library(data.table)
library(fields)
library(grid)
library(lubridate)
library(magick)
library(ncdf4)
library(sf)
library(terra)
library(vioplot)


### load buffer data
setwd("~/R_projects/US_Shrimp_ONG/data")

yrs <- 2015:2019
for(i in 1:length(yrs)){
  yr <- yrs[i]
  tmp <- read.csv(paste0(yr,'_eff_ves_dist4.csv'))
  assign(paste0('eff_',yr), tmp)
}

effort <- rbind(eff_2015,eff_2016,eff_2017,eff_2018,eff_2019)
effort$year <- as.numeric(sort(paste(rep(yrs,126))))

eff_pr_agg <- aggregate(cbind(effort$effort_pr,effort$effort_hrs),
                        by = list(effort$dist_km, effort$statzone),
                        median, na.rm=T) |>
  setNames(c('dist_km','statzone','effort_pr','effort_hrs'))

ves_pr_agg <- aggregate(cbind(effort$vessel_pr,effort$vessel_nu),
                        by = list(effort$dist_km, effort$statzone),
                        median, na.rm=T) |>
  setNames(c('dist_km','statzone','vessel_pr','vessel_nu'))


pts <- rep(c(21,24,23),3)
cols <- cmocean('thermal')(length(pts))


### load encounter metrics
setwd("~/R_projects/US_Shrimp_ONG")
# load('./data/eff_out_1000.rdata')
# load('./data/eff_out_1000-2.rdata')
load('./data/eff_out_1000-4.rdata')
load('./data/rpts_10000-2.rdata')

setwd("~/R_projects/US_Shrimp_ONG/data")
load('horizon.RData')

rpts_out_all <- tmp
rm(tmp)
rpts_out_all <- rpts_out_all[c(1,7:14,2:6)]
names(rpts_out_all)

dist_km <- c(.5,1,2,3,5,7.5,10,12.5,15,17.5,20,25,30,40)
units(dist_km) <- 'km'

rpts_out <- matrix(NA, 14, 9)
for(i in 1:14){
  rigs_p <- with(rpts_out_all[[i]],
                 rbind(as.vector(sapply(rig_p, function(x) sum(x[2:length(x)]))),
                       as.vector(sapply(rig_p, function(x) x[1]))))
  rpts_out[i,] <- rigs_p[1,]
}


eff_out <- array(NA, c(5, 14, 9))
for(h in 2015:2019){
  # h=2017
  ind <- grep(h, names(eff_out_all))
  eff_t <- eff_out_all[ind]
  eff_t <- eff_t[c(1:2,7,10,13:14,3:6,8:9,11:12)]

  # eff_out <- matrix(NA, 14, 9)
  for(i in 1:14){
    rigs_p <- with(eff_t[[i]],
                   rbind(as.vector(sapply(rig_p, function(x) sum(x[2:length(x)]))),
                         as.vector(sapply(rig_p, function(x) x[1]))))
    eff_out[h-2014, i,] <- rigs_p[1,]
  }

}
eff_out[is.na(eff_out)] <- 0
eff_out_m <- apply(eff_out, c(2,3), median, na.rm=T)

s_i <- (eff_out_m - rpts_out)
brks <- seq(-round(max(abs(s_i),na.rm=T),digits = 1),
            round(max(abs(s_i),na.rm=T),digits = 1),
            length.out = 61)


pts <- rep(c(21,24,23),3)
cols <- cmocean('thermal')(length(pts))

### load mini map png
setwd("~/R_projects/US_Shrimp_ONG/figures")
mini_sz <- image_read('mini_SZ_plot.png')

setwd("~/R_projects/US_Shrimp_ONG/figures")

png('buffers_encounter.png', width = 16, height = 11, units = 'in', res = 300)
par(mfrow=c(2,2),mar=c(4,4,1,1))

plot(eff_pr_agg$dist_km, eff_pr_agg$effort_pr, typ = 'n',
     xlab='', ylab='', las = 1)
# rect(0, 0, hor, 1, col = 'gray80', border = NA)
rect(hor, 0, hor_twothirds, 1, col = 'gray80', border = NA)
rect(hor_twothirds, 0, hor_top, 1, col = 'gray50', border = NA)
rect(hor_top, 0, 40, 1, col = 'gray30', border = NA)
grid()
for(j in 13:21){
  with(subset(eff_pr_agg, statzone==j),{
    points(dist_km, effort_pr,
           typ = 'l', pch = pts[j-12], col = cols[j-12], lwd = 2)
    points(dist_km, effort_pr,
           typ = 'p', pch = pts[j-12], bg = cols[j-12], cex = 1.5)
  }
  )
}
# legend('topleft', legend=c(13:21), pch = pts, pt.bg = cols, bty = 'n')
mtext('Cumulative proportion of effort', 2, 2.5)
mtext('a)',side=3, adj=0, font=2)

plot(0,0,xlim=c(0,40),ylim=c(0,.45), typ = 'n',
     xlab='', ylab='', las = 1)
rect(hor, 0, hor_twothirds, 1e5, col = 'gray80', border = NA)
rect(hor_twothirds, 0, hor_top, 1e5, col = 'gray50', border = NA)
rect(hor_top, 0, 40, 1e5, col = 'gray30', border = NA)
grid()
text(hor, .45, '\n horizon', srt = 90, adj = 1)
text(hor_twothirds, .45, '\n partially visible', srt = 90, adj = 1, col = 'white')
text(hor_top, .45, '\n not visible', srt = 90, adj = 1, col = 'white')
for(i in 13:21){
  tmp <- subset(eff_pr_agg, statzone==i)
  points(tmp$dist_km,
         c(tmp$effort_hrs[1], diff(tmp$effort_hrs))/sum(c(tmp$effort_hrs[1], diff(tmp$effort_hrs))),
         typ = 'l',  col = cols[i-12], lwd = 2)
  points(tmp$dist_km,
         c(tmp$effort_hrs[1], diff(tmp$effort_hrs))/sum(c(tmp$effort_hrs[1], diff(tmp$effort_hrs))),
         typ = 'p',  pch = pts[i-12], bg = cols[i-12], cex = 1.5)
}
# legend('topleft', legend=c(paste(13:16,'LA',sep='-'),
#                            paste(17,'LA/TX',sep='-'),
#                            paste(18:21,'TX',sep='-')),
#        pch = pts, pt.bg = cols, bty = 'n')
mtext('Proportion of effort', 2, 2.5)
mtext('b)',side=3, adj=0, font=2)


plot(ves_pr_agg$dist_km, ves_pr_agg$vessel_pr, typ = 'n',
     xlab='', ylab='', las = 1)
# rect(0, 0, hor, 1, col = 'gray80', border = NA)
rect(hor, 0, hor_twothirds, 1, col = 'gray80', border = NA)
rect(hor_twothirds, 0, hor_top, 1, col = 'gray50', border = NA)
rect(hor_top, 0, 40, 1, col = 'gray30', border = NA)
grid()
text(hor, .45, '\n horizon', srt = 90, adj = 1)
text(hor_twothirds, .8, '\n partially visible', srt = 90, adj = 1, col = 'white')
text(hor_top, .8, '\n not visible', srt = 90, adj = 1, col = 'white')
for(j in 13:21){
  with(subset(ves_pr_agg, statzone==j),{
    points(dist_km, vessel_pr,
           typ = 'l', pch = pts[j-12], col = cols[j-12], lwd = 2)
    points(dist_km, vessel_pr,
           typ = 'p', pch = pts[j-12], bg = cols[j-12], cex = 1.5)
  }
  )
}
mtext('Cumulative proportion of vessels', 2, 2.5)
mtext('Distance from rig (km)',1, 2.5)
mtext('c)',side=3, adj=0, font=2)
# rasterImage(mini_sz, 30, 0, 40, .5)

plot(as.numeric(dist_km), rep(0, length(dist_km)),
     ylim = range(s_i), typ = 'n',
     xlab = '', ylab = '', las = 1)
rect(hor, -1, hor_twothirds, 1, col = 'gray80', border = NA)
rect(hor_twothirds, -1, hor_top, 1, col = 'gray50', border = NA)
rect(hor_top, -1, 40, 1, col = 'gray30', border = NA)
grid()
rug(x = seq(-.25,.05,.1), side = 2, ticksize = -.02, lwd = 1.5)
# mtext(2017)
text(hor, -.45, '\n horizon', srt = 90, adj = 1)
text(hor_twothirds, -.45, '\n partially visible', srt = 90, adj = 1, col = 'white')
text(hor_top, -.45, '\n not visible', srt = 90, adj = 1, col = 'white')
for(i in 1:9){
  points(dist_km, s_i[,i], typ ='l', lwd = 4)
  points(dist_km, s_i[,i], typ ='l', lwd = 2, col = cols[i])
  points(dist_km, s_i[,i], typ = 'p', pch = pts[i], bg = cols[i], cex = 1.5)
}
abline(h = 0, lty = 5, lwd = 2)
legend('bottomleft', legend=c(paste(13:16,'LA',sep='-'),
                              paste(17,'LA/TX',sep='-'),
                              paste(18:21,'TX',sep='-')),
       pch = pts, pt.bg = cols, bty = 'n', cex = 1.25)
# legend('bottomleft', legend = c(13:21), pch = pts, pt.bg = cols, bty = 'n')
mtext('Rig Avoidance [-] vs. Attraction [+]', 2, 2.5)
mtext('Distance from observation (km)',1, 2.5)
mtext('d)',side=3, adj=0, font=2)

grid.raster(mini_sz, x = .38, y = .17, width = .2)

dev.off()
